Mega Update
Add omdb and search commands and remove out of data message files and initialization code. Update config.json and package.json.
