+++ package.json
@@ -1,6 +1,6 @@
 {
    "name": "deldbot",
-   "version": "1.0.0",
+   "version": "2.0.0",
    "description": "Deldbot heeds the call of the Sons of Lanarchy",
    "main": "bot.js",
    "author": "Derek Elder",
