--- src/util.js
@@ -1,3 +1,0 @@
-function totalPages (number) {
-  return Math.ceil(number / 10)
-}
